% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxillary.R
\name{create.negbin}
\alias{create.negbin}
\alias{create.poisson}
\title{Create negative binomial}
\usage{
create.negbin(r)

create.poisson()
}
\arguments{
\item{r}{}
}
\description{
Create negative binomial

Create Poisson model
}
\examples{
Afkt <- create.negbin(4.5)
#simfd.ed(..., ed.fct = Afkt, ...) ## Use with simfd.ed

Afkt <- create.poisson()
#simfd.ed(..., ed.fct = Afkt, ...) ## Use with simfd.ed
}
