% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postanalysis.R
\name{align_data_plot}
\alias{align_data_plot}
\title{Plot aligned data}
\usage{
align_data_plot(y, t, w, warp_fct, coords, type = "l", ...)
}
\arguments{
\item{y, t}{data}

\item{w}{warp. If missing, no alignment is done.}

\item{warp_fct}{}

\item{coords}{Which coordinates of trajectories? E.g. 1 for fist column.}

\item{type}{plot type}

\item{...}{Further arguments to be passed to plot and points.}
}
\description{
Plot aligned data
}
\details{
Much improvement needs to be done on this function.
}
