% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arm_data_example}
\alias{arm_data_example}
\title{Arm movement data}
\format{An object of class \code{list} of length 2.}
\usage{
arm_data_example
}
\description{
A small selection of arm movement data avaliable from \url{github.com/larslau/Bochum_movement_data}
This file contains the 10 curves from experimental setup 8, individual 1. 
The curves have been processed to simm.fda format.
}
\details{
\describe{
\item{t} {List of time points. Time points are placed equidistantly from 0 to 1.}
\item{y} {List of observations. Each element is a m_i \code{\times} 3 matrix.}
}
}
\keyword{datasets}
