% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kovarianser.R
\name{Brown}
\alias{Brown}
\title{Brownian motion and Brownian bridge}
\usage{
Brown(t, par = c(tau = 1, noise = 1), motion = FALSE)
}
\arguments{
\item{t}{time points}

\item{par}{Two-dimensional. Variance of BB/BM and variance of noise.}

\item{motion}{logical. Brownian bridge or brownian motion}
}
\description{
Brownian motion and Brownian bridge
}

