% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_functions.R
\name{default_warp_cov}
\alias{default_warp_cov}
\title{Brownian Bridge warp covariance}
\usage{
default_warp_cov(default.value)
}
\arguments{
\item{default.value}{}
}
\description{
This is called default covariance function as this is typically used
}

