% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kovMat.R
\name{dekomp}
\alias{dekomp}
\title{dekomp}
\usage{
dekomp(Y)
}
\arguments{
\item{Y}{Positivite definite matrix}
}
\value{
Y^{1/2}
}
\description{
Positive definite square root of matrix
}
\details{
Does not check if Y is positive definite
}

