% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kovarianser.R
\name{diag_covariance}
\alias{diag_covariance}
\title{Diagonal covariance}
\usage{
diag_covariance(t, par)
}
\arguments{
\item{t}{time points}

\item{par}{parameters. One parameter for each dimension}
}
\value{
Covariance matrix
}
\description{
This is the simplest covariance function.
}
\examples{
# Three-dimensional curve:
amp_cov <- diag_covariance
amp_cov_par <- c(1,1,1)
ppMulti(y,t, etc...)
}
