% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kovMat.R
\name{kovMat}
\alias{kovMat}
\title{Dynamical covariance structure}
\usage{
kovMat(t, a, tw, timefunction, stack = TRUE, noise = 0, ...)
}
\arguments{
\item{t}{evalutation points}

\item{a}{list of matrices}

\item{tw}{Anchor points. Must be in increasing order}

\item{timefunction}{Temporal covariance structure}

\item{stack}{How should stack the observations. TRUE corresponds to the implementation of ppMUlti, while FALSE is the alternative ordering.}

\item{noise}{Noise. Either a single number or a vector of length K*length(t)}

\item{...}{Parameters passed to timefunction}
}
\value{
A positive definite covariance matrix
}
\description{
The machinery for handling dynamical covariance structure
}
\references{
See pdf simm.fda for introduction and of details of dynamical covariance structure
}
