% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp-multi_ed.R
\name{likelihood.ed}
\alias{likelihood.ed}
\title{Likelihood in ED models}
\usage{
likelihood.ed(param, param.w, r, Zis, A.diff2, amp_cov, warp_cov, t, tw,
  pr = FALSE)
}
\arguments{
\item{param}{Amplitude parameters}

\item{param.w}{Warp covariance parameters}

\item{r, Zis}{Residual vectors/matrices}

\item{A.diff2}{2nd derivative for exponential family}

\item{amp_cov}{}

\item{warp_cov}{}

\item{t}{Observation time points}

\item{tw}{}

\item{pr}{Print option}
}
\value{
A number
}
\description{
Likelihood in ED models
}
