% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp-multi_1.R
\name{likelihood}
\alias{likelihood}
\title{Linearised likelihood function}
\usage{
likelihood(param, param.w, r, Zis, amp_cov, warp_cov, t, tw, sig = F,
  pr = F)
}
\arguments{
\item{param}{}

\item{param.w}{}

\item{r}{}

\item{Zis}{}

\item{amp_cov}{}

\item{warp_cov}{}

\item{t}{}

\item{tw}{}

\item{sig}{}

\item{pr}{Print value?}
}
\value{
-2*logL(parameters)
}
\description{
Linearised likelihood function
}

