% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxillary.R
\name{like.nowarp}
\alias{like.nowarp}
\alias{likelihood}
\alias{like.par}
\alias{likelihood.lap}
\title{Linearised likelihood functions.}
\usage{
like.nowarp(param, r, amp_cov, t, sig = FALSE, pr = FALSE, ...)

likelihood(param, r, amp_cov, t, param.w, Zis, warp_cov, tw, sig = FALSE,
  pr = FALSE, w = NULL, parallel = NULL)

like.par(param, r, amp_cov, t, param.w, Zis, warp_cov, tw, sig = FALSE,
  pr = FALSE, w = NULL, parallel = NULL)

like.nowarp(param, r, amp_cov, t, sig = FALSE, pr = FALSE, ...)

likelihood.lap(param, r, amp_cov, t, param.w, Zis, warp_cov, tw, sig = FALSE,
  pr = FALSE, w, parallel = FALSE)
}
\arguments{
\item{param}{Amplitude parameters}

\item{r, Zis}{Residual vectors/matrices}

\item{amp_cov}{Amplitude covariance function}

\item{t}{Observation time points}

\item{sig}{Return sigma or likelihood?}

\item{pr}{Print option. Only used for debugging.}

\item{param.w}{Warp parameters}

\item{warp_cov}{Warp covariance function}

\item{tw}{}
}
\value{
-2*logL(parameters)
}
\description{
Linearised likelihood functions.
}
\details{
\code{likelihood} is the default option (linearised likehood). \code{like.par} performs parallelized likelihood. 
\code{likelihood.lap} uses Laplace approximation. If w is the optimal posterior, then Laplace approximation and linearisation are the same, 
in general lik_laplace <= lik_lin.
\code{like.nowarp} is used when there is no warping used (and thus no linearisation).
}
\seealso{
\link{ppMulti}
}
