% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_functions.R
\name{make_fourier_basis}
\alias{make_fourier_basis}
\title{Generate Fourier basis function}
\usage{
make_fourier_basis(endpoints, order)
}
\arguments{
\item{endpoints}{Left and right endpoints.}

\item{order}{Order of harmonics}
}
\description{
Generates a fourier basis 'basis function'
}
\details{
The evaluated function has intercept at its first index , followed first by cosine coefficients and then by sine coefficients, both in increasing order.
}
\seealso{
\link{make_basis_fct}
}

