% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kovarianser.R
\name{Maternk}
\alias{Maternk}
\title{Matern covariance}
\usage{
Maternk(t, par = c(range = 1, smooth = 2, noise = 1))
}
\arguments{
\item{t}{time points}

\item{par}{parameters}
}
\value{
Covariance matrix for Matern covariance
}
\description{
Matern covariance
}
\examples{
t <- seq(0,1 , 0.01)
Maternk(t, c(0.1, 2, 1))
}

