% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kovarianser.R
\name{mvMatern}
\alias{mvMatern}
\title{Multivariate Matern covariance}
\usage{
mvMatern(t, range = 1, smooth = 2, noise = 1, sig)
}
\arguments{
\item{t}{}

\item{range, smooth}{range and smoothness parameters}

\item{smooth}{}

\item{sig}{Marginal covariance. K x K matrix}

\item{noise.}{noise for different coordinates. Must be a vector of length K.}
}
\description{
Multivariate Matern covariance
}
\details{
noise and sig should fit in dimension
}
\examples{

}
