% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kovarianser.R
\name{mvOUproces}
\alias{mvOUproces}
\title{Multivariate OU process}
\usage{
mvOUproces(t, lambda = 1, sig, noise = 0)
}
\arguments{
\item{t}{time points}

\item{lambda}{drift parameter.}

\item{sig}{Marginal covariance. K x K matrix}

\item{noise}{}
}
\description{
Multivariate OU process
}

