% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{nlm.bound.xx}
\alias{nlm.bound.xx}
\title{Bounded optimization using nlm}
\usage{
nlm.bound.xx(fct, p, lower, upper, symmetric = FALSE,
  flagInfinity = TRUE, init = FALSE, ...)
}
\arguments{
\item{fct}{Function to be optimized over. One vector argument.}

\item{p}{Starting values}

\item{lower}{lower bounds for paramters.}

\item{upper}{upper bounds for parameters.}

\item{flagInfinity}{Avoid Inf/Inf expressions (Only used if symmetric = FALSE). If FALSE NaN values may appear.}

\item{...}{Further arguments to nlm}
}
\value{

}
\description{
Performs bounded (box-constrained) non-linear minimization using the fast but unconstrained nlm optimizer. 
This function has in many cases been faster than box-constrained optimization using \link{optim}.
}
\details{
nlm.bound doesn't check validity of bounds, but does check if starting values satisfy the constraints.
If any starting value is not strictly between its prescribed bounds, it is adjusted with a warning. 

symmetric = FALSE uses a logit transformation while symmetric = TRUE uses an asymmetrical logarithm.
symmetric =  TRUE is expected to be a better choice in most cases.
}
\seealso{
\link{nlm}
}
