% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kovarianser.R
\name{OUproces}
\alias{OUproces}
\title{Ornstein-Uhlenbeck process}
\usage{
OUproces(t, par = c(lambda = 1, noise = 0))
}
\arguments{
\item{t}{time points}

\item{par}{vector with drift and noise terms.}
}
\value{
Covariance matrix for OU process
}
\description{
Ornstein-Uhlenbeck process
}
\details{
Note that the marginal variance (expect for noise) is assumed to be 1.
}
