% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kovMat.R
\name{poly.bridge}
\alias{poly.bridge}
\title{Polynomial bridge}
\usage{
poly.bridge(tmin, tmax, coef)
}
\arguments{
\item{tmin, }{tmax Max and min values}

\item{coef}{Coefficients, i.e. a marginal variance given by coef[1] + t*(1-t)*coef[2]}
}
\value{
If done correctly, a 'bridge structure covariance', see
}
\description{
Polynomial bridge
}
\details{
Polynomial bridge cannot be used as a timefunction. It is recommended to combine this with timefunction.bridge
or something like it to ensure the right structure.
}
\seealso{
timefunction.bridge
}
