% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kovMat.R
\name{poly.larger}
\alias{poly.larger}
\title{Polynomial 'bridge' of arbitrary order}
\usage{
poly.larger(tmin, tmax, coef)
}
\arguments{
\item{tmin, tmax}{Max and min values}

\item{coef}{Polynomial coefficients for bridge, in increasing order.}
}
\description{
Polynomial 'bridge' of arbitrary order
}
\details{
Generalise \code{poly.bridge} to arbitrary order.
}
\seealso{
\link{poly.bro}, timefunction.bridge
}
