% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kovMat.R
\name{poly.Matern}
\alias{poly.Matern}
\title{Matern covariance and bridge}
\usage{
poly.Matern(s, t, range, smooth, koef)
}
\arguments{
\item{s, t}{}

\item{t}{}

\item{range}{}

\item{smooth}{}

\item{koef}{}
}
\description{
Matern covariance and bridge
}
\details{
This is a specific implementation for Matern covariance.
}
\examples{
f <- function(s, t, params) poly.Matern(s, t, params[1], params[2], params[3:4] )

a1 <- diag(c(1.5, 0.1, 0.7))
a2 <- a1 + 2

ti <- seq(0,1, 0.02)

kovMat(ti, list(a1, a2), c(0,1), f, params = c(0.5, 2, 1, 1), noise = 0.1)

}
\seealso{
\link{timefunction.bridge}
}

