% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp-multi.R
\name{posterior.lik}
\alias{posterior.lik}
\title{Posterior likelihood}
\usage{
posterior.lik(w, warp_fct, t, y, basis_fct, c, Sinv, Cinv)
}
\arguments{
\item{w}{warp values}

\item{warp_fct}{}

\item{t}{time points}

\item{y}{observations}

\item{basis_fct}{}

\item{c}{matrix of spline coefficients}

\item{Sinv}{Precision matrix for amplitude}

\item{Cinv}{Precision matrix for w}
}
\description{
Calculates the posterior likelihood for a single curve.
}

