% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp-multi_2.R
\name{pp.multi.like}
\alias{pp.multi.like}
\title{ppmultilike}
\usage{
pp.multi.like(
  y,
  t,
  basis_fct,
  warp_fct,
  w,
  amp_cov = NULL,
  warp_cov = NULL,
  amp_cov_par = NULL,
  design = NULL,
  parallel.lik = FALSE,
  use.laplace = FALSE
)
}
\description{
ppmultilike
}
