% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp-multi-c.R
\name{ppMulti.em}
\alias{ppMulti.em}
\title{ppMulti with update steps using EM algorithm}
\usage{
ppMulti.em(y, t, basis_fct, warp_fct, amp_cov = NULL, warp_cov = NULL,
  iter = c(5, 5), parallel = list(n_cores = 1, parallel_likelihood = FALSE),
  amp_cov_par = NULL, paramMax = rep(T, length(amp_cov_par)), w.c.p = T,
  like.alt = F, like_optim_control = list(), pr = T, design = NULL,
  win = T, inner_control = list(), save_temp = NULL, w0 = NULL)
}
\arguments{
\item{y}{}

\item{t}{}

\item{basis_fct}{}

\item{warp_fct}{}

\item{amp_cov}{}

\item{warp_cov}{}

\item{iter}{}

\item{parallel}{}

\item{amp_cov_par}{}

\item{paramMax}{}

\item{w.c.p}{}

\item{like.alt}{}

\item{like_optim_control}{}

\item{pr}{}

\item{design}{May not be null}

\item{win}{}

\item{inner_control}{}

\item{save_temp}{}

\item{w0}{}
}
\description{
This is like ppMulti, but uses one step of the EM algorithm to update spline coefficients in the linearized model.
Theoretically better than ppMulti, but slower and the differences should be neglible.
Requires spam package
}
\details{
There has been less check on this function, so I cannot guarantee that it will behave as well as ppMulti.
Requires \code{spam} package, used for faster calculations for sparse matrices.
}
\seealso{
\link{ppMulti}
}
