% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp-multi_2.R
\name{ppMulti.ny}
\alias{ppMulti.ny}
\title{Simultaneous inference for misaligned multivarariate functional data}
\usage{
ppMulti.ny(y, t, basis_fct, warp_fct, amp_cov = NULL, warp_cov = NULL,
  iter = c(5, 5), w0 = NULL, use.nlm = c(FALSE, FALSE),
  functional = NULL, amp_cov_par = NULL, paramMax = rep(TRUE,
  length(amp_cov_par)), parallel.lik = c(FALSE, FALSE),
  warp_opt = TRUE, like_optim_control = list(), pr = FALSE,
  design = NULL, inner_parallel = FALSE, save_temp = NULL)
}
\arguments{
\item{parameters}{Similar to \link{ppMulti}}
}
\value{
@see \link{ppMulti}
}
\description{
Similar to ppMulti, but uses Laplace approximation and not linearization. 
When w^0 is the max posterior logL_lin = logL_lap; in general logL_lin <= logL_lap. 
This function calculates likelihod slightly faster than ppMulti; otherwise I have opinion.
}
