% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funktiontest.R
\name{ppMulti.test}
\alias{ppMulti.test}
\title{Test}
\usage{
ppMulti.test(y, t, basis_fct, warp_fct, amp_cov = NULL,
  warp_cov = NULL, iter = c(5, 5), use.nlm = c(FALSE, FALSE),
  functional = NULL, amp_cov_par = NULL, paramMax = rep(T,
  length(amp_cov_par)), warp_opt = TRUE, parallel.lik = FALSE,
  like_optim_control = list(), pr = TRUE, design = NULL,
  inner_parallel = c(TRUE, TRUE), spammed = T, save_temp = NULL,
  w0 = NULL, c0 = NULL)
}
\description{
Test
}
