% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postanalysis.R
\name{predict_warps}
\alias{predict_warps}
\title{Evaluated warp functions}
\usage{
predict_warps(w, warp_fct, seq.length = 101, plot = TRUE, tout, ...)
}
\arguments{
\item{w}{warp values}

\item{seq.length}{number of time points used in the evaluation.}

\item{plot}{Plot the result? Uses matplot.}

\item{tout}{Evaluation points. If missing, equidistant evaluation on [0,1] according to seq.length is used in the evaluation}

\item{...}{arguments to be passed to matplot}
}
\value{
A matrix of dim seq.length x n
}
\description{
Returns and (optionally) plots predicted warps.
}
\examples{
 
\donttest{

tw <- c(0.25, 0.5, 0.75)
wf <- <- make_warp_fct(type="smooth", tw=tw)

<run estimation on data set and return object res>

predict_warps(res$w, wf, plot = T, type="l", main = "Predicted warp functions")

}

}
