% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_functions.R
\name{shift_covariance}
\alias{shift_covariance}
\title{Simple one-parameter covariance for pure shift}
\usage{
shift_covariance(default.value)
}
\arguments{
\item{default.value}{Default value}
}
\description{
Simple one-parameter covariance for pure shift
}

