% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp-multi_ed.R
\name{simfd.ed}
\alias{simfd.ed}
\title{Simultaneous inference for misaligned functional data in an exponential family setting}
\usage{
simfd.ed(y, t, basis_fct, warp_fct, ed_fct, amp_cov = NULL, warp_cov = NULL,
  iter = c(5, 5), w0 = NULL, u0 = NULL, use.nlm = c(FALSE, FALSE),
  suppressLik = FALSE, amp_cov_par = NULL, paramMax = rep(TRUE,
  length(amp_cov_par)), warp_opt = TRUE, like_optim_control = list(),
  design = NULL, inner_parallel = FALSE, save_temp = NULL)
}
\arguments{
\item{y}{List of observations. NAs allowed}

\item{t}{List of corresponding time points. NAs not allowed}

\item{basis_fct}{Basis function for spline}

\item{warp_fct}{Warp function}

\item{ed_fct}{Function defining the exponential family. Must have attribute diff2. See \link{Afkt} for examples.}

\item{amp_cov}{Amplitude covariance function. Must be on the form \code{function(t, param)}}

\item{warp_cov}{Warp covariance function. Must be on the form \code{function(t, param)}}

\item{iter}{two-dimensional integer of maximal number of outer iterations &
maximal number of inner iterations per outer iteration.}

\item{w0}{Starting values for predicted warps. Should only be used if you have results from a previous run.}

\item{u0}{Starting values for predicted trajectories. Should only be used if you have results from a previous run.}

\item{use.nlm}{Use \code{nlm} instead of \code{optim} for optimization? First index for outer loop, second index for inner loop.}

\item{suppressLik}{Suppress if likelihood has increased}

\item{amp_cov_par}{Starting values for amplitude covariance parameters. There are no defaults.}

\item{paramMax}{Logical vector. Which amplitude parameters to optimise over? Defaults to all parameters.
May be overwritten by supplying control parameters.}

\item{warp_opt}{If FALSE, warp covariance parameters are kept fixed.}

\item{like_optim_control}{List of control options for optimization in outer loop. See \link{ppMulti} for details.}

\item{design}{Design for the experiments. Should be given as a list of one-dimensional vectors or as a design matrix.}

\item{inner_parallel}{Should the inner optimization be done in parallel?}

\item{save_temp}{Save estimates after each outer iteration? NULL or the file path.}
}
\value{
A list of estimates and predictions of w and u
}
\description{
This function applies the simm.fda methodology in a setting of certain exponential models (e.g. poisson) followiong the approach outlined in article[..]
}
\details{
\code{sim.fd} can only handle one-dimensional functional data. It should NOT be applied to Gaussian data, unless sigma^2 is known and fixed beforehand.
}
\seealso{
\link{ppMulti}
}
