% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kovMat.R
\name{OUtid}
\alias{OUtid}
\alias{BMtid}
\alias{Materntid}
\title{Temporal covariance structures}
\usage{
OUtid(s, t, lambda)

BMtid(s, t, bridge = TRUE)

Materntid(s, t, range, smooth)
}
\arguments{
\item{s, t}{time values}

\item{lambda}{drift}

\item{bridge}{Brownian bridge or Brownian motion?}

\item{range}{range parameter}

\item{smooth}{smoothness parameter}
}
\value{
A vector. Proper use is together with \code{outer} or \code{kovMat}.
}
\description{
Temporal covariance structures

Matern covariance structure: 
2^(smooth-1)/gamma(smooth) * range^smooth ^ K_smooth(|s-t|/range)
}
\details{
Time covariance functions should not be called directly. Either use as part of kovMat or call using outer.
}
\examples{

}
