% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kovMat.R
\name{timefunction.bridge}
\alias{timefunction.bridge}
\title{Timefunction and bridge}
\usage{
timefunction.bridge(s, t, coef, timefunction, ...)
}
\arguments{
\item{s, t}{}

\item{coef}{Coefficients for bridge}

\item{timefunction}{A timefunction}

\item{...}{Parameters passed to timefunction}
}
\value{
A time function structure.
}
\description{
Creates a 'bridged' time function, i.e. a combination of an 'ordinary' time function and a bridge covariance. 
As the result is a time function, it can be used in conjunction with kovMat.
}
\examples{

f <- function(s, t, param) timefunction.bridge(s, t, coef = c(param[1],param[2]), OUtid, lambda = param[3])

ti <- seq(0, 1, 0.02)

outer(ti , ti, f, param = c(0.2, 1, 0.7)) 
outer(ti , ti, f, param = c(1, 0, 0.7)) 
# Try make a contour plot to see the differences


}

