% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warping_functions.R
\name{w.shift}
\alias{w.shift}
\title{Create shift-&-warp function}
\usage{
w.shift(warp.fct, maxShift)
}
\arguments{
\item{warp.fct}{A warp function without shift}

\item{maxShift}{Maximal shift allowed. More than \code{\pm} 0.3 would rarely make sense.}
}
\value{
A warp function with one extra parameter
}
\description{
Create shift-&-warp function
}
