% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_functions.R
\name{warp_and_shift_cov}
\alias{warp_and_shift_cov}
\title{Brownian Bridge warp covariance + shift parameter}
\usage{
warp_and_shift_cov(default.values)
}
\arguments{
\item{default.values}{c(shift variance, Brownian Bridge variance)}
}
\value{
Covariance function
}
\description{
Combines default_warp_cov and a shift parameter. Shift and warp are assumed to be independent.
}
\seealso{
\link{default_warp_cov}, \link{w.shift}
}
