% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxillary.R
\name{likelihood}
\alias{likelihood}
\alias{like.par}
\title{Linearised likelihood function}
\usage{
likelihood(param, param.w, r, Zis, amp_cov, warp_cov, t, tw, sig = FALSE,
  pr = FALSE)

like.par(param, param.w, r, Zis, amp_cov, warp_cov, t, tw, sig = FALSE,
  pr = FALSE)
}
\arguments{
\item{param}{Amplitude parameters}

\item{param.w}{Warp parameters}

\item{r, Zis}{Residual vectors/matrices}

\item{amp_cov}{Amplitude covariance function}

\item{warp_cov}{Warp covariance function}

\item{t}{Observation time points}

\item{tw}{}

\item{sig}{Return sigma or likelihood?}

\item{pr}{Print option. Only use for debugging.}
}
\value{
-2*logL(parameters)
}
\description{
Linearised likelihood function
}
\details{
like.par performs parallelized likelihood
}
