% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxillary.R
\name{likelihood}
\alias{likelihood}
\alias{like.par}
\alias{like.nowarp}
\title{Linearised likelihood functions.}
\usage{
likelihood(param, r, amp_cov, t, param.w, Zis, warp_cov, tw, sig = FALSE,
  pr = FALSE, w = NULL)

like.par(param, r, amp_cov, t, param.w, Zis, warp_cov, tw, sig = FALSE,
  pr = FALSE, w = NULL)

like.nowarp(param, r, amp_cov, t, sig = FALSE, pr = FALSE, ...)
}
\arguments{
\item{param}{Amplitude parameters}

\item{r, Zis}{Residual vectors/matrices}

\item{amp_cov}{Amplitude covariance function}

\item{t}{Observation time points}

\item{param.w}{Warp parameters}

\item{warp_cov}{Warp covariance function}

\item{tw}{}

\item{sig}{Return sigma or likelihood?}

\item{pr}{Print option. Only used for debugging.}
}
\value{
-2*logL(parameters)
}
\description{
Linearised likelihood functions.
}
\details{
\code{likelihood} is the default option. \code{like.par} performs parallelized likelihood. 
\code{like.nowarp} is used when there is no warping used (and thus no linearisation).
}
\seealso{
\link{ppMulti}
}
