% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{nlm.bound}
\alias{nlm.bound}
\title{Bounded optimization using nlm}
\usage{
nlm.bound(fct, p, lower, upper, ...)
}
\arguments{
\item{fct}{Function to be optimized over. One vector argument.}

\item{p}{Starting values}

\item{lower}{lower bounds for paramters.}

\item{upper}{upper bounds for parameters.}

\item{...}{Further arguments to nlm}
}
\value{

}
\description{
Performs bounded (box-constrained) non-linear minimization using the fast but unconstrained \code{nlm} optimizer. 
This function is in many cases been faster than box-constrained optimization using \link{optim}.
}
\details{
nlm.bound doesn't check validity of the bounds, but does check if starting values satisfy the constraints.
If any starting value is not strictly between its prescribed bounds, it is adjusted with a warning.
}
\seealso{
\link{nlm}
}
