% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_functions.R
\name{order2_covariance}
\alias{order2_covariance}
\title{Second-order derivative covariance}
\usage{
order2_covariance(default.value)
}
\arguments{
\item{default.value}{Default value}
}
\value{
Covariance function
}
\description{
Second-order derivative covariance
}
