% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp-multi-c.R
\name{pavpop_returner_zogr}
\alias{pavpop_returner_zogr}
\title{Returns linearization stuff}
\usage{
pavpop_returner_zogr(y, t, basis_fct, warp_fct, amp_cov = NULL,
  warp_cov = NULL, iter = c(5, 5), parallel = list(n_cores = 1,
  parallel_likelihood = FALSE), use_warp_gradient = FALSE,
  amp_cov_par = NULL, w0 = NULL, pr = T, design = NULL, win = T,
  inner_control = list())
}
\arguments{
\item{y}{}

\item{t}{}

\item{basis_fct}{}

\item{warp_fct}{}

\item{amp_cov}{}

\item{warp_cov}{}

\item{iter}{}

\item{parallel}{}

\item{use_warp_gradient}{}

\item{amp_cov_par}{}

\item{w0}{}

\item{pr}{}

\item{design}{}

\item{win}{}

\item{inner_control}{}
}
\value{

}
\description{
Returns linearization stuff
}
