% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp-multi.R
\name{ppMulti}
\alias{ppMulti}
\alias{simm.fda}
\title{Simulataneous inference for Misaligned Multivarariate functional data}
\usage{
ppMulti(y, t, basis_fct, warp_fct, amp_cov = NULL, warp_cov = NULL,
  iter = c(5, 5), parallel = list(n_cores = 1, parallel_likelihood = FALSE),
  amp_cov_par = NULL, paramMax = rep(T, length(amp_cov_par)), w.c.p = T,
  like.alt = F, like_optim_control = list(), pr = F, design = NULL,
  win = F, save_temp = NULL)
}
\arguments{
\item{y}{list of observations in matrix form. NAs allowed}

\item{t}{list of corresponding time points}

\item{basis_fct}{Basis fct.}

\item{warp_fct}{Warp fct.}

\item{amp_cov}{Amplitude covariance. Must be on the form function(t, param)}

\item{warp_cov}{Warp covariance}

\item{iter}{two-dimensional integer of maximal number ofouter iterations 
and maxiamal number of inner iterations per outer iteration.}

\item{parallel}{}

\item{amp_cov_par}{Starting values for amplitude covariance parameters. There are no defaults.}

\item{paramMax}{logical vecter. Which amplitude parameters to optimise over? Defaults to all parameters.}

\item{w.c.p}{Do not change}

\item{like.alt}{Do not change either}

\item{like_optim_control}{List of control options for optimization in outer loop. See details}

\item{pr}{print option}

\item{design}{Design for the experiments. Should be given as a list of one-dimensional vectors.}

\item{win}{}

\item{save_temp}{Save estimates after each outer iteration? FALSE, NULL or the file path.}
}
\value{
A list of estimates
}
\description{
Simulataneous inference for Misaligned Multivarariate functional data
}
\details{
ppMulti returns error if applied on one-dimensional functional data.

Control parameters
lower, upper, method, ndeps, maxit will be sent to optim. See optim for more details. 
The first entries of lower/upper correspond to warp parameters, while the rest corresponds to
amplitude parameters. ppMulti does match upper/lower with corresponding entries in amp_cov, which is important when not all parameters are maximized over. 
This is for consistency with randomCycle and optimRule.

randomCycle and optimRule are two ways of optimizing on only a subset of the parameters at a time. TBD: descriptions of these.
}
\examples{
\\link{example}

\donttest{
# Data originates from http://mocap.cs.cmu.edu/

## Data is provided as
# MCD.data y-values
# MCD.time time points

# Make basis function
bf <- make_basis_fct(seq(0, 1, len=32)[2:31], intercept=T, boundary = c(0, 1))

## Make warp function
tw <- seq(0, 1, length = 5)[2:4] ## anchor points for hyman spline
wf.noshift <- make_warp_fct(type="smooth", tw=tw)

## warp function with shift
wf <- w.shift(wf.noshift, 0.25)

warp_cov <- warp_and_shift_cov(c(0.5, 1))


## For this example we assume no amplitude cross-correlation, otherwise ...
# Matern covariance; smoothness parameter 2, unknown range.

wrapper <- function(t, par) {
  mvMatern(t, par[1], 2, rep(1,3), diag(par[2:4]))
}

lower <- c(1e-4, 1e-4,rep(1e-4, 4))
upper <- rep(1e5, 6)

# Note this will take a long time:
mcd.res <-ppMulti(MCD.data, MCD.time, bf, wf, amp_cov = wrapper, warp_cov, amp_cov_par = c(0.2, rep(100, 3)), pr = F, paramMax = rep(T, 4),
                  like_optim_control = list(lower = lower, upper = upper), win = F, iter = c(10, 10))
}

}
\seealso{
Important details can be found in simm-fda-short-desc.pdf
}

